requirements.txt
matplotlib==3.3.4
numpy==1.20.2
Pillow==8.0.1
opencv-python==4.5.1.48
seaborn==0.11.1
torch==1.8.0
torchvision==0.9.0
tqdm==4.60.0